more pack functionality, bug fixes

+++ MapAC/DatLoader/BinaryReaderExtensions.cs
@@ -68,7 +68,7 @@ namespace MapAC.DatLoader
                 return (uint)(knownType + (higher | lower));
             }
             */
-            writer.Write(value - knownType);
+            writer.Write((ushort)(value - knownType));
         }
 
         /// <summary>
